requirements.txt

python-multipart==0.0.6
pillow==10.1.0
tensorflow-cpu==2.18.0
torch==2.1.1
torchvision==0.16.1
opencv-python==4.8.1.78
numpy==2.0.2
requests==2.31.0
python-dotenv==1.0.0
ultralytics==8.0.196
google-generativeai==0.3.2
protobuf==3.20.3
gradio==4.31.2